backend/requirements.txt
# Django Core
django==5.2
uvicorn==0.30.0

# Django REST Framework
djangorestframework==3.15.2
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.4.0
djangorestframework-camel-case==1.4.2

# Django Extensions
django-cors-headers==4.6.0
django-anymail[mailgun]==12.0
whitenoise==6.7.0

# Security & Monitoring
sentry-sdk[django]==1.40.6
argon2-cffi==23.1.0

# Testing & Development
pytest==8.1.0
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0
ruff==0.1.13
mypy==1.8.0

# Production Database
psycopg[binary]==3.2.3
